day7

+++ WS/address-book-rest-api/src/main/java/com/cts/adb/service/ContactServiceImpl.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.cts.adb.entities.Contact;
+import com.cts.adb.exceptions.ContactNotFoundException;
 import com.cts.adb.repos.ContactRepo;
 
 @Service
@@ -20,13 +21,20 @@ public class ContactServiceImpl implements ContactService {
 	}
 
 	@Override
-	public void deleteById(int contactId) {
+	public void deleteById(int contactId) throws ContactNotFoundException{
+		if(!contactRepo.existsById(contactId)) {
+			throw new ContactNotFoundException("Contact#"+contactId+" doesn't exist to delete!");
+		}
 		contactRepo.deleteById(contactId);
 	}
 
 	@Override
-	public Contact getById(int contactId) {
-		return contactRepo.findById(contactId).orElse(null);
+	public Contact getById(int contactId) throws ContactNotFoundException {
+		Contact contact = contactRepo.findById(contactId).orElse(null);
+		if(contact==null) {
+			throw new ContactNotFoundException("Contact#"+contactId+" doesn't exist!");
+		}
+		return contact;
 	}
 
 	@Override
@@ -35,8 +43,20 @@ public class ContactServiceImpl implements ContactService {
 	}
 	
 	@Override
-	public Contact getByMobile(String mmbielNumber) {
-		return contactRepo.findByMobileNumber(mmbielNumber).orElse(null);
+	public Contact getByMobile(String mmbielNumber) throws ContactNotFoundException {
+		Contact contact = contactRepo.findByMobileNumber(mmbielNumber).orElse(null);
+		if(contact==null) {
+			throw new ContactNotFoundException("Contact with mobile number"+ mmbielNumber +" doesnt exist!");
+		}
+		return contact;
+	}
+
+	@Override
+	public Contact update(Contact contact) throws ContactNotFoundException {
+		if(!contactRepo.existsById(contact.getContactId())) {
+			throw new ContactNotFoundException("Contact#"+contact.getContactId()+" doesnt exist to update!");
+		}
+		return contactRepo.save(contact);
 	}
 }
